+++ source/modifiers/paralysis.js
@@ -2,8 +2,8 @@ const { ModifierTemplate } = require("../classes");
 
 module.exports = new ModifierTemplate("Paralysis",
 	"Gain 1 Stagger instead of losing it between rounds for @{stackCount} rounds.",
-	null,
+	"<:Paralysis:1266521422663123025>",
 	false,
 	true,
 	1
-);
+).setInverse("Agility");